+++ src/App/auth.js
@@ -1,15 +1,21 @@
 import {
   signInWithEmailAndPassword,
   AuthErrorCodes,
+  getAuth, signOut,
+  signInWithPopup,
+  GoogleAuthProvider,
 } from 'firebase/auth';
 // import { GoogleAuthProvider } from “firebase/auth”;
 // eslint-disable-next-line import/no-cycle
 import {
   auth,
 } from './firebase';
-import { goTo } from './routes';
+import {
+  goTo,
+} from './routes';
 // eslint-disable-next-line max-len
 // exportando función de login
+// const auth = getAuth();
 export const login = (email, password, changeHashFn) => {
   signInWithEmailAndPassword(auth, email, password).then((userCredential) => {
     // Signed in
@@ -31,3 +37,44 @@ export const login = (email, password, changeHashFn) => {
       }
     });
 };
+
+// Login Con Google
+const provider = new GoogleAuthProvider();
+export const loginWithGoogle = () => {
+  signInWithPopup(auth, provider)
+    .then((result) => {
+      // This gives you a Google Access Token. You can use it to access the Google API.
+      const credential = GoogleAuthProvider.credentialFromResult(result);
+      const token = credential.accessToken;
+      // The signed-in user info.
+      const user = result.user;
+      console.log(user);
+      console.log('SESION INICIADA');
+      goTo('/wall');
+      // ...
+    })
+    .catch((error) => {
+      // Handle Errors here.
+      const errorCode = error.code;
+      const errorMessage = error.message;
+      // The email of the user's account used.
+      alert('El email ingresado ya se encuentra registrado');
+      const email = error.customData.email;
+      // The AuthCredential type that was used.
+      alert('El email ingresado ya se encuentra registrado');
+      const credential = GoogleAuthProvider.credentialFromError(error);
+      // ...
+    });
+};
+
+// CERRAR SESIÓN
+export const CerrarSesión = () => {
+  signOut(auth).then(() => {
+    goTo('/home');
+  // Sign-out successful.
+  })
+    .catch((error) => {
+      alert('An error happened');
+    });
+};
+
